requirements.txt
Django==5.1.2
pillow==11.0.0
crispy-tailwind==1.0.3
django-formtools==2.5.1
requests==2.32.3
django-environ==0.11.2